Refactored example to have core Run() in separate package, new window in "demo"
