+++ internal/example/doc.go
@@ -1,5 +1,5 @@
-// Package demo contains the core logic of the demo.
+// Package example contains the core logic of the demo.
 // The Run() function demonstrates how a typical application would create a UI.
 // The function herein is not concerned about technology-specific things, such as
 // which abstraction library or which drawing interface is used.
-package demo
+package example
